# swagger.yaml
openapi: 3.0.0
info:
  title: Geofence API
  version: 1.0.0
servers:
  - url: https://stg-nemo.mahindralastmilemobility.com:2083/api/v1/fms
  - url: https://stg-nemo.mahindralastmilemobility.com:2053/api/v1/fms
  - url: http://localhost:10004/api/v1/fms
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
paths:
...
  /report/alert:
    post:
      tags:
        - FMS > Report
      summary: Get Geo Alert Report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              example:
                fleetid: "2c360155-e1cb-40cb-80f3-f222eaffedb6"
                ruleids: ["11532bc8-9668-448e-89a6-5734b9721b08"]
                vinnos: ["VIN123456789"]
                starttime: 1752670110
                endtime: 1752756510
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Geo alert report
          content:
            application/json: {}
  /report/trip:
    post:
      tags:
        - FMS > Report
      summary: Get Geo Trip Report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              example:
                fleetid: "2c360155-e1cb-40cb-80f3-f222eaffedb6"
                ruleids: ["11532bc8-9668-448e-89a6-5734b9721b08"]
                vinnos: ["VIN123456789"]
                starttime: 1752670110
                endtime: 1752756510
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Geo trip report
          content:
            application/json: {}